requirements.txt
torch==2.6.0
torchvision>=0.17.0
numpy>=1.26.0
opencv-python>=4.9.0
matplotlib>=3.8.0
Pillow>=10.2.0
tqdm>=4.66.0